Move leader schedule into separate module
